Implemented a Game Class
This game class organizes the code by isolating the different parts of the main game loop into different methods
