+++ src/ffnn.hpp
@@ -3,12 +3,7 @@
 
 #include <iostream>
 #include <fstream>
-#include <chrono>
-#include <random>
 
-#include <stddef.h>
-
-#include <cinttypes>
 #include <cmath>
 
 #include <Eigen/Core>
