+++ src/gamestats.cpp
@@ -1,4 +1,3 @@
-#include <algorithm>
 #include <string>
 
 #include "gamestats.hpp"
