+++ src/pipe.cpp
@@ -1,5 +1,4 @@
 #include "pipe.hpp"
-#include "config.hpp"
 
 ////////////////////////// Pipe //////////////////////////
