+++ src/pipe.hpp
@@ -4,9 +4,6 @@
 #include <iostream>
 #include <array>
 
-#include <cstddef>
-#include <cstdint>
-
 #include <SFML/Graphics/Color.hpp>
 #include <SFML/Graphics/Drawable.hpp>
 #include <SFML/Graphics/RectangleShape.hpp>
